+++ examples/backup-example/main.tf
@@ -1,5 +1,5 @@
 terraform {
-  required_version = "~>0.12.15"
+  required_version = "> 0.12.26"
 
   required_providers {
     azurerm = {